requirements.txt
Django==2.2.6
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
djoser==2.0.3
Pillow==6.2.0
psycopg2==2.8.2
requests==2.22.0
